[S05P13C105-107][BE] refactor: Team Id Integer로 변경

+++ backend/src/main/java/com/web/curation/team/Team.java
@@ -25,7 +25,7 @@ public class Team {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "team_id")
-    private long teamId;
+    private Integer teamId;
 
     private String name;
